+++ src/main/java/net/chaos/chaosmod/biomes/CustomBiomeNether.java
@@ -78,9 +78,11 @@ public class CustomBiomeNether extends BiomeHell {
 	
 	@Override
 	public int getGrassColorAtPos(BlockPos pos) {
-		// return 0x7ca4f7;
-		// return 0x1a60f0;
 		return 0xff0000;
-		// return ColorizerGrass.getGrassColor(0.0F, 0.5F);
+	}
+	
+	@Override
+	public int getFoliageColorAtPos(BlockPos pos) {
+		return 0xff0000; // TODO : tweak to make leaves stand out over grass
 	}
 }